+++ extensions/amp-a4a/0.1/refresh-manager.js
@@ -316,7 +316,7 @@ export class RefreshManager {
             this.config_.visiblePercentageMin).unobserve(this.element_);
         this.a4a_.refresh(() => this.initiateRefreshCycle());
         resolve(true);
-      }, /** @type {number} */ (2000));
+      }, /** @type {number} */ (this.refreshInterval_));
     });
   }
